refactor(types): extract bug tests to sibling file
